03-json-and-nlp-in-pgsql/lecture-notes/references/psycopg/workspace/pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.0",
    "psycopg[binary,pool]>=3.2.9",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]
